refactor: Move Score component & useAnimatedScore hook to their own files
